Repository design partten

+++ database/migrations/2020_12_30_065300_create_categories_table.php
@@ -6,11 +6,6 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateCategoriesTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::create('categories', function (Blueprint $table) {
@@ -28,17 +23,12 @@ class CreateCategoriesTable extends Migration
             $table->integer('updated_by')       ->unsigned()->nullable();
             $table->integer('deleted_by')       ->unsigned()->nullable();
 
-            $table->unsignedInteger('parent_id')->default(0);
-
+            $table->unsignedBigInteger('parent_id')->default(0);
+            
             $table->timestamps();
         });
     }
-
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
+    
     public function down()
     {
         Schema::dropIfExists('categories');
